+++ src/components/Hand/index.tsx
@@ -5,10 +5,13 @@ type HandProps = {
   shouldDisplayHand: boolean;
 };
 
-const Hand = ({ shouldDisplayHand }: HandProps): JSX.Element =>
-  <Style
-    className={`hand${shouldDisplayHand ? '' : ' hidden'}`}
-    onClick={(event: React.MouseEvent<HTMLInputElement>): void => event.stopPropagation()}>
-  </Style>;
+const Hand = ({ shouldDisplayHand }: Readonly<HandProps>): JSX.Element => {
+  return (
+    <Style
+      className={`hand${shouldDisplayHand ? '' : ' hidden'}`}
+      onClick={(event: React.MouseEvent<HTMLInputElement>): void => event.stopPropagation()}>
+    </Style>
+  );
+};
 
 export default Hand;