+++ docs/changelog.md
@@ -35,6 +35,7 @@ Ends August 17, 2020
 - update buildspec so that testing closes onces finished.
 - prevent Hand component from hiding when it's clicked.
 - hide Hand component if it's already displaying and a Card compnent is clicked.
+- Refactor App component to pass down setShouldDisplayHand with Context.
 
 ## Sprint 1
 Ended August 3, 2020